📝 add xml docs & rename `Map` to `SgbMap` to avoid namespace collisions (#4)

+++ src/Spillgebees.Blazor.Map/Models/Layers/Path.cs
@@ -5,8 +5,8 @@ public interface IPath
     public bool Stroke { get; init; }
     public string? StrokeColor { get; init; }
     public int? StrokeWeight { get; init; }
-    public int? StrokeOpacity { get; init; }
+    public double? StrokeOpacity { get; init; }
     public bool Fill { get; init; }
     public string? FillColor { get; init; }
-    public int? FillOpacity { get; init; }
+    public double? FillOpacity { get; init; }
 }
